strip down shell view
