+++ frontend/src/components/panels/item-plugin-panel.tsx
@@ -1,0 +1,23 @@
+import { PluginType, usePluginState } from '@app/../../core/src';
+import { PluginContent } from '../organisms/tile-action';
+
+export const ItemPluginPanel = () => {
+    const ui = usePluginState();
+    const itemPluginStates = (ui || [])
+        .filter((p) => p.config.type === PluginType.ITEM)
+        .flatMap((p) => p.state.components.flatMap((c) => c.content));
+
+    return (
+        <>
+            {itemPluginStates.length > 0 && (
+                <div className="tile-actions action">
+                    <div className="controls">
+                        {itemPluginStates.map((content, idx) =>
+                            content ? <PluginContent key={idx} content={content} /> : undefined
+                        )}
+                    </div>
+                </div>
+            )}
+        </>
+    );
+};
